+++ satella/opentracing/__init__.py
@@ -1,3 +1,3 @@
 from .trace import trace_future, trace_function
-
-__all__ = ['trace_future', 'trace_function']
+from .exceptions import trace_exception
+__all__ = ['trace_future', 'trace_function', 'trace_exception']