+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * added `intify`
 * added `satella.coding.typing`
+* added `satella.opentracing.trace_exception`